requirements.txt
matplotlib==3.5.0
numpy==1.18.1
pandas==0.24.2
Pillow==8.4.0
scikit_learn==1.0.2
seaborn==0.10.0
streamlit==1.2.0
textblob==0.15.3
